requirements.txt
feedparser==5.2.1
flask==1.1.1
wtforms==2.2.1
python-dotenv==0.12.0
flask-sqlalchemy==2.4.1
flask-mail==0.9.1
flask-login==0.5.0
flask_apscheduler==1.11.0
flask-wtf==0.14.3
itsdangerous==1.1.0
python-memcached==1.59